minor fixes in ssa

+++ secureserviceaccount/src/main/java/com/autodesk/aps/secureserviceaccount/model/ExchangeJwtToken.java
@@ -4,7 +4,7 @@
  * The APS Platform contains an expanding collection of web service components that can be used with Autodesk cloud-based products or your own technologies. Take advantage of Autodesk’s expertise in design and engineering.
  *
  * Secure Service Account
- * Operations to manage Service accounts and keys.   A service account is an identity that an application can use to make requests to other services without a user authorizing the requests. A service account is identified by a unique email address and has an Autodesk ID.  A service account has one or more private keys. A private key is generated through an asymmetric cryptography algorithm; the paired public key is stored by Autodesk Identity.  An application can use a service account's private key to generate a JWT token. The JWT token provides proof of implicit authentication and authorization for this service account; an application can exchange it for a three-legged access token for the service service.
+ * Operations to manage Service accounts and keys.  A service account is an identity that an application can use to make requests to other services without a user authorizing the requests. A service account is identified by a unique email address and has an Autodesk ID.  A service account has one or more private keys. A private key is generated through an asymmetric cryptography algorithm; the paired public key is stored by Autodesk Identity.  An application can use a service account's private key to generate a JWT token. The JWT token provides proof of implicit authentication and authorization for this service account; an application can exchange it for a three-legged access token for the service account.
  *
  * Contact: aps.help@autodesk.com
  *
@@ -24,12 +24,16 @@
 package com.autodesk.aps.secureserviceaccount.model;
 
 import java.util.Objects;
-import java.util.Arrays;
-import io.swagger.v3.oas.annotations.media.Schema;
-import com.fasterxml.jackson.annotation.JsonTypeName;
-import java.math.BigDecimal;
+
+import com.fasterxml.jackson.annotation.JsonCreator;
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
-import com.fasterxml.jackson.annotation.*;
+import com.fasterxml.jackson.annotation.JsonTypeName;
+import com.fasterxml.jackson.annotation.JsonValue;
+
+import io.swagger.v3.oas.annotations.media.Schema;
 
 /**
  * The response for exchange JWT
@@ -85,7 +89,7 @@ public class ExchangeJwtToken {
   private TokenTypeEnum tokenType;
 
   public static final String JSON_PROPERTY_EXPIRES_IN = "expires_in";
-  private BigDecimal expiresIn;
+  private Integer expiresIn;
 
   public static final String JSON_PROPERTY_EXPIRES_AT = "expires_at";
   private Long expiresAt;
@@ -145,7 +149,7 @@ public class ExchangeJwtToken {
   }
 
 
-  public ExchangeJwtToken expiresIn(BigDecimal expiresIn) {
+  public ExchangeJwtToken expiresIn(Integer expiresIn) {
     
     this.expiresIn = expiresIn;
     return this;
@@ -160,14 +164,14 @@ public class ExchangeJwtToken {
   @JsonProperty(JSON_PROPERTY_EXPIRES_IN)
   @JsonInclude(value = JsonInclude.Include.NON_NULL)
 
-  public BigDecimal getExpiresIn() {
+  public Integer getExpiresIn() {
     return expiresIn;
   }
 
 
   @JsonProperty(JSON_PROPERTY_EXPIRES_IN)
   @JsonInclude(value = JsonInclude.Include.NON_NULL)
-  public void setExpiresIn(BigDecimal expiresIn) {
+  public void setExpiresIn(Integer expiresIn) {
     this.expiresIn = expiresIn;
   }
